Updated DENM and CAM Encoding, now the ASN.1 DENM and CAM are standard compliant (still some issue in DENM)

+++ src/automotive/model/appSample.h
@@ -9,7 +9,7 @@ typedef struct cam_field_t
   std::pair<double,double> pos;
   double speed;
   double acceleration;
-  double angle;
+  int angle;
 } m_cam_field;
 
 typedef struct denm_field_t
